Add generated tracing registration coverage (#366)
Document the test subscriber's callback contracts and generate capture
histories across filter levels, thread ordering, and child subscriber
state. Keep the deterministic child-first regression as the concrete
proof of the reported race.

Split the focused test module to keep Rust source files within the
repository size limit, and record the expanded verification scope.
